Confirm forward and inverse solution of a unit square, and changed some terminology in mechanicsproblem.py
The forward and inverse elongation of a unit square was solved using
the MechanicsProblem object. Also, the keys 'lame1' and 'lame2' used
in the config dictionary have been changed to 'lambda' and 'mu',
respectively. The idea is for these variables to take on the values
appropriate for the constitutive equation being used.